Prepare for packaging - updated pyproject.toml - updated typing in selector_element

+++ llm_generation_server/elements/selector_element.py
@@ -122,7 +122,7 @@ class MinMaxSubElement(SelectorSubElement):
     """
 
     def __init__(
-        self, sample_min: int, sample_max: int, text: str, step_size: float = 1.0
+        self, sample_min: float, sample_max: float, text: str, step_size: float = 1.0
     ):
         super().__init__(subtype="min_max", text=text)
         if sample_min >= sample_max:
@@ -130,7 +130,7 @@ class MinMaxSubElement(SelectorSubElement):
                 f"sample_min ({sample_min}) should be bigger than sample_max "
                 + f"({sample_max})"
             )
-        self._selected: int = sample_min
+        self._selected: float = sample_min
         self._min = sample_min
         self._max = sample_max
         self._step_size = step_size
